Merges EXASOL_FS into EXASOL dialect #1495 Fixes the Bracket Problem within SCRIPTS #1479
