+++ src/sqlfluff/dialects/exasol_keywords.py
@@ -851,6 +851,7 @@ UNRESERVED_KEYWORDS = [
     "GROUP_TEMP_DB_RAM_LIMIT",
     "JDBC",
     "PRECEDENCE",
+    "ROWCOUNT",
     "STEP",
     "SESSION_TEMP_DB_RAM_LIMIT",
     "USER_TEMP_DB_RAM_LIMIT",
