+++ test/fixtures/parser/exasol/CreateAdapterScriptStatement.sql
@@ -1,2 +1,3 @@
 CREATE JAVA ADAPTER SCRIPT my_script AS
 %jar hive_jdbc_adapter.jar;
+/
