+++ test/fixtures/parser/exasol/CreateAdapterScriptStatement.yml
