+++ test/fixtures/parser/exasol/CreateFunctionStatement.sql
