+++ test/fixtures/parser/exasol/CreateFunctionStatement.yml
