+++ test/fixtures/parser/exasol/CreateLuaScriptBracket.sql
@@ -1,0 +1,6 @@
+CREATE OR REPLACE LUA SCRIPT BRACKET.SCRIPT_EXAMPLE RETURNS ROWCOUNT AS
+    local _stmt = [[SOME ASSIGNMENT WITH OPEN BRACKET ( ]]
+    x = 1
+    local _stmt = _stmt .. [[ ) ]]
+return 1
+/
