+++ test/fixtures/parser/exasol/CreatePythonScalarScript.sql
