+++ test/fixtures/parser/exasol/CreatePythonScalarScript.yml
