+++ test/fixtures/parser/exasol/CreateScriptingLuaScriptStatement1.yml
