+++ test/fixtures/parser/exasol/CreateScriptingLuaScriptStatement2.sql
