+++ test/fixtures/parser/exasol/CreateScriptingLuaScriptStatement3.sql
