+++ test/fixtures/parser/exasol/CreateUDFScriptStatement1.sql
