+++ test/fixtures/parser/exasol/CreateUDFScriptStatement1.yml
