+++ test/fixtures/parser/exasol/CreateUDFScriptStatement2.sql
