+++ test/fixtures/parser/exasol/CreateUDFScriptStatement3.yml
