+++ test/fixtures/parser/exasol/CreateUDFScriptStatement4.yml
