+++ test/fixtures/parser/exasol/CreateUDFScriptStatement5.sql
